\name{smcUtils-deprecated}
\alias{smcUtils-deprecated}
\title{Deprecated Functions in Package smcUtils}
\description{These functions are provided for compatilbity with older versions of smcUtils and may be defunct as soon as the next release.}
\seealso{\code{\link{Deprecated}},\code{\link{renormalize.weights-deprecated}}}
